tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/events/014.html

<!doctype html>
<title>WebSockets: setting event handlers to null</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
var events = ['onclose', 'onopen', 'onerror', 'onmessage'];
for (var i = 0; i < events.length; ++i) {
  test(function() {
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/empty-message');
    var foo = function() {}
    ws[events[i]] = foo;
    assert_equals(ws[events[i]], foo, events[i]);
    ws[events[i]] = null;
    assert_equals(ws[events[i]], null, events[i]);
  }, "Setting event handlers to null " + events[i]);
};
</script>
